requirements.txt
discord.py
pillow
python-dotenv
flask